Collection and SchemasReader are working!
SchemasReader is our class that reads the Schemas on app/schemas.
This is great because with this we can create a hash schema of our
little DB.

This also has other functions like getting the file names and
collection names DIRECTLY FROM THE YAMLs!!

It's helpful to have that because now we can dynamically create the
instance arrays of the models collections.

Collection is a singleton class that holds all the information,
you can just do:

Collection.instance.<model>

For instance if we have a schema and model persons, we can just do:

Collection.instance.persons

And we'll get the array with the persons objects.