Add collection specs & refactor for it to work
In order for the specs to work I had to change the Collection
define_methods on file load to the initializer by using the send
method and calling define_method there.

Also when running all the specs the persons_spec failed because
it already was instanciated before, so now I just force the method
and instance variable of the Collection with an allow.

That doesn't matter now because it was already tested on the
Collection specs.