ID-able Models!
Models/resources are now IDable. This means that everytime you create
a new resource it's gonna create an ID for it.

You can either pass it as a parameter when creating it, or don't and
wait for it to be autogenerated.

The calculated id takes the highest ID and adds 1 to that. So if the
highest ID is 33 the new ID will be 34.

I.E.

manuel = ::Models::Person.create('Manuel', 23)
manuel.id == 23

OR

ed = ::Models::Person.create('Eduardo')
ed.id = 24